Grouped REST tests by common endpoint
