+++ rest/tests/test_misc.py
@@ -1,0 +1,50 @@
+"""
+Unit tests for REST api
+"""
+from __future__ import unicode_literals
+
+from rest_framework.status import (
+    HTTP_404_NOT_FOUND,
+)
+
+from rest.tests.base import (
+    RESTTestCase,
+    API_BASE,
+)
+
+
+class TestMisc(RESTTestCase):
+    """
+    REST test
+    """
+
+    def test_root(self):
+        """
+        Test root of API
+        """
+        resp = self.client.get(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.post(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.head(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.patch(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.put(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.options(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+
+        self.logout()
+        resp = self.client.get(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.post(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.head(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.patch(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.put(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
+        resp = self.client.options(API_BASE)
+        self.assertEqual(HTTP_404_NOT_FOUND, resp.status_code)
